refactor(interactive_checkout): break out into another action
